Revive the nuked Chinese localizations, and add a hack for Korean.
The initial Korean localization attempt apparently nuked the existing
Chinese text, and to make sure our Chinese friends don't get angry at
us (and to be as least intrusive as possible) this adds a nasty little
hack that manipualtes the DOM upon load to only leave the Korean
localized elements left when ?l=ko-kr is appended to the URL. Otherwise
the document should be exactly the same as the original copy.

But we should really try to fix this so we can do proper i18n.